test/web-platform-tests/to-upstream/css/cssom-view/Element-getClientRects.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.getClientRects</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-element-getclientrects">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_object_equals(el.getClientRects(), []);

}, "getClientRects should return empty array for an element with no associated CSS layout box");
</script>
